requirements.txt
kafka-python==2.0.2
aiohttp==3.8.3
redis==4.3.4
python-dotenv==0.20.0
web3==5.31.0